Extract system_domain module in domain service
No logical changes
